Separating routes and state interfaces
